Some fixes and API for randomized Cat image now works.

+++ app/src/main/java/com/example/projekt/Cat.kt
@@ -10,8 +10,8 @@ class Cat {
     @SerializedName("webpurl")
     var myWebpurl: String = ""
     @SerializedName("x")
-    var myX: Int = 0
+    var myX: Double = 0.0
     @SerializedName("y")
-    var MyY: Int = 0
+    var MyY: Double = 0.0
 
 }
